Pulling data from local json files.
